Append SEO toolkit button in the same way on page builder contents
